Add options to perform ordering while ensuring unique results
This increases the complexity of the Query class even further,
and it's in dire need of refactoring to make it understandable
again.

However, I will hold off on this for a little bit longer to implement
a few extra features first, so that those will also be included in the
refactoring.