First step in major refactor of query handling
This changes the PGQuery class to first "flatten" the nested query into a set
of entities and filters, and only create the actual database query and joins
in the final step. Splitting this will make the code more understandable as
well as extensible.

Some functionality is temporarily broken by this, such as using Blobs as
main target as well as sorting and "final wheres".